Related to #36, #37 Add validation to ConfirmsModal Add Header component Add ConnectionModal component Add jazzicon to address
