+++ app/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@ethersproject/providers": "^5.0.23",
+    "@metamask/jazzicon": "^2.0.0",
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
     "@web3-react/ledger-connector": "^6.1.9",
