+++ app/pages/_app.tsx
@@ -1,13 +1,10 @@
 import '../styles/globals.css';
 import Head from 'next/head';
+import { Header } from 'components/Header';
 import { WithModal } from 'components/Modal';
 import { WithToast, Toast } from 'components/Toast';
 import { Web3Provider } from '@ethersproject/providers';
 import { Web3ReactProvider } from '@web3-react/core';
-import { ChainId } from 'components/ChainId';
-import { BlockNumber } from 'components/BlockNumber';
-import { NetworkName } from 'components/NetworkName';
-import { Connection } from 'components/Connection';
 function getLibrary(provider: any): Web3Provider {
   const library = new Web3Provider(provider);
   library.pollingInterval = 12000;
@@ -31,11 +28,8 @@ const MyApp = ({
         <WithModal>
           <WithToast>
             <div className="container mx-auto px-4 pb-4">
-              <div className="flex justify-end mt-2">
-                <ChainId />
-                <BlockNumber />
-                <NetworkName />
-                <Connection />
+              <div className="flex items-stretch justify-end mt-2">
+                <Header />
               </div>
               <div className="mt-4">
                 <Toast />
